requirements.txt
openai-whisper
ffmpeg-python
pyyaml
PyQt6
